Save progress- this is really buggy
It either reverses the order of the object properties or fails to
recognzie commas and just keeps adding characters to currentToken. Or it
just doesn't add the properties at all. I really have no idea why any of
this is happening. I 'fixed' the no props added bug, but I don't
understand why the fix works.